lxd/storage/drivers: Don't probe filesystem on snapshot deactivate
Since the refcount name for snapshots and their parents is now the same,
we no longer need to explicitly check the parent's refcount to make sure
that deactivate is valid; deactivate simply won't be called for in-use
snapshots.

Based on the git history I suspect that the filesystem probing here was
left over from before mount refcounts were added; it's unclear to me
what additional situations it catches that refcounts do not.

Signed-off-by: Wesley Hershberger <wesley.hershberger@canonical.com>